First commit. \o/
So far everything's in CoffeeScript, but I'm eyeballing the possibility of an ES version in the near future.

Most data is stored on IPFS, with the server using its own FS as a sort of buffer while posts "fill up" over the course of presumebly-days.

HTML is inline in the script for now. The ONLY time the www server will feed from the FS directly is for requests to .well-known/* - '..' requests are overzealously blocked. Any uncaught error or promise will bring the server down. (that's on purpose)

Have fun. :j